refactor(core): move ParseResult to separate file
