+++ packages/api-elements/lib/parse-result.js
@@ -8,66 +8,15 @@
  */
 
 const apiDescription = require('./api-description');
+const parseResult = require('./elements/parse-result');
 
 const namespace = (options) => {
+  parseResult(options.base);
+
   const minim = options.base;
   const { Element } = minim;
   const StringElement = minim.getElementClass('string');
   const ArrayElement = minim.getElementClass('array');
-
-  /**
-   * @class ParseResult
-   *
-   * @param {Array} content
-   * @param meta
-   * @param attributes
-   *
-   * @extends ArrayElement
-   */
-  class ParseResult extends ArrayElement {
-    constructor(...args) {
-      super(...args);
-      this.element = 'parseResult';
-    }
-
-    /**
-     * @name api
-     * @type Category
-     * @memberof ParseResult.prototype
-     */
-    get api() {
-      return this.children.filter(item => item.classes.contains('api')).first;
-    }
-
-    /**
-     * @name annotations
-     * @type ArraySlice
-     * @memberof ParseResult.prototype
-     */
-    get annotations() {
-      return this.children.filter(item => item.element === 'annotation');
-    }
-
-    /**
-     * @name warnings
-     * @type ArraySlice
-     * @memberof ParseResult.prototype
-     */
-    get warnings() {
-      return this.children
-        .filter(item => item.element === 'annotation' && item.classes.contains('warning'));
-    }
-
-    /**
-     * @name errors
-     * @type ArraySlice
-     * @memberof ParseResult.prototype
-     */
-    get errors() {
-      return this.children
-        .filter(item => item.element === 'annotation' && item.classes.contains('error'));
-    }
-  }
 
   /**
    * @class Annotation
@@ -137,7 +86,6 @@ const namespace = (options) => {
 
   minim
     .use(apiDescription)
-    .register('parseResult', ParseResult)
     .register('annotation', Annotation)
     .register('sourceMap', SourceMap);
 };
